Files for search bar
As of now, if you search for a source that matches something in the database exactly, it will redirect to samplesearch.html and populate it with that information. If it's not in the database, it will still redirect to samplesearch.html, but will say it wasn't found.

Bugs: It doesn't query the database properly if you hit enter instead of the "Search Database" button. I tried a whole bunch of Stack Overflow solutions that I'm pretty sure should have worked to make an Enter press act exactly like a button click but I still couldn't get it to work. Feel free to offer some solutions, but it's probably not an important bug to fix for the scope of this project.